Added AssetTracker.Client
Added AssetTracker.Client project to the solution which is a simple ASP.NET Core MVC Web Application.  Application brings up all the Assets in a list and you can select the details which will show the detail view.

Also added a new migration to add OrganizationId to the Statuses and Types tables so these lookup lists will now be associated with an Organization.